refactor: json-rpc-provider package
