+++ packages/json-rpc-provider/src/index.ts
@@ -1,0 +1,3 @@
+export { exponentialBackoff } from './exponential-backoff';
+export { JsonRpcProvider } from './json-rpc-provider';
+export { fetchJson } from './web';
